_examples: branch example improvements
